% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impulse_DE_fin.R
\name{calc_impulse}
\alias{calc_impulse}
\title{Impulse model value prediction}
\usage{
calc_impulse(theta, timepoints)
}
\arguments{
\item{theta}{numerical vector of impulse parameters with the order
beta, h0, h1, h2, t1, t2.}

\item{timepoints}{numercial vector of time point(s).}
}
\value{
The predicted impulse model values for the given time point(s).
}
\description{
Calculates impulse model values for given timepoints and predicted
impulse parameters.
}
\examples{
#' theta vector in the order beta, h0, h1, h2, t1, t2
theta <- c(9.9, 14.7, 17.0, 16.9, -0.1, 37.0)
#' time points
timepoints <- c(0, 2, 4, 6, 8, 18, 24, 32, 48, 72)
#' calculate impulse values
impulse_values <- calc_impulse(theta, timepoints)
}
\author{
Jil Sander
}
\references{
Chechik, G. and Koller, D. (2009) Timing of Gene Expression
Responses to Envi-ronmental Changes. J. Comput. Biol., 16, 279-290.
}
\seealso{
\code{\link{impulse_DE}}, \code{\link{plot_impulse}}.
}

